Move game to separate fragment
